requirements.txt
torch
torchvision
timm
albumentations
gradio
pandas
numpy
opencv-python-headless
tqdm
Pillow
